+++ test/phpunit/PlayersByTeam/Response/developer.json.NE.header.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @link      https://github.com/gridiron-guru/FantasyDataAPI for the canonical source repository
+ * @copyright Copyright (c) 2014 Robert Gunnar Johnson Jr.
+ * @license   http://opensource.org/licenses/Apache-2.0
+ * @package   FantasyDataAPI
+ */
+
+$headers = [];
+$headers[] = 'HTTP/1.1 200 OK';
+$headers[] = 'Cache-Control: must-revalidate, max-age=0, private';
+$headers[] = 'Content-Length: 372425';
+$headers[] = 'Content-Type: application/json; charset=utf-8';
+$headers[] = 'Expires: Fri, 30 May 2014 02:12:19 GMT';
+$headers[] = 'Last-Modified: Fri, 30 May 2014 02:08:29 GMT';
+$headers[] = 'Vary: Accept,Accept-Charset';
+$headers[] = 'X-Powered-By: ASP.NET';
+$headers[] = 'X-Mashape-Billing: Call Limit=1';
+$headers[] = 'Date: Fri, 30 May 2014 02:12:19 GMT';
+
+return $headers;
